[test] Separate integration and unit tests (#156)
* Separate integration and unit tests

* New tests/unit and tests/integration directories

* Add missing H2 machines to conftest

* Move offline backend tests to unit/
