+++ .github/workflows/build_and_test.yml
@@ -7,6 +7,7 @@ on:
       - develop
   push:
     branches:
+      - main
       - develop
       - 'wheel/**'
       - 'runci/**'
@@ -17,6 +18,13 @@ on:
   schedule:
     # 04:00 every Tuesday morning
     - cron: '0 4 * * 2'
+  workflow_dispatch:
+    inputs:
+      run-integration-tests:
+        description: "Please type 'yes' to run also the integration tests (default 'no')."
+        required: true
+        default: "no"
+
 
 env:
   PYTKET_REMOTE_QUANTINUUM_USERNAME: ${{ secrets.PYTKET_REMOTE_QUANTINUUM_USERNAME }}
@@ -48,27 +56,101 @@ jobs:
       uses: actions/setup-python@v4
       with:
         python-version: '3.10'
-    - name: Build and test including remote checks (3.10) mypy
+    - name: Build and test (3.10) mypy
       shell: bash
-      if:  (matrix.os == 'macos-12') && (github.event_name == 'push' || (github.event_name == 'pull_request' &&  github.event.pull_request.head.repo.full_name == github.repository) || github.event_name == 'release' || github.event_name == 'schedule' )
+      if: |
+        matrix.os == 'macos-12' &&
+        (
+          github.event_name == 'push' ||
+          (
+            github.event_name == 'pull_request' &&
+            github.event.pull_request.head.repo.full_name == github.repository
+          ) ||
+          github.event_name == 'release' ||
+          (
+            github.event_name == 'workflow_dispatch' &&
+            github.event.inputs.run-integration-tests == 'no'
+          )
+        )
       run: |
         ./.github/workflows/build-test mypy
+    - name: Build and test including integration (3.10) mypy
+      shell: bash
+      if: |
+        matrix.os == 'macos-12' &&
+        (
+          (
+            github.event_name == 'push' &&
+            (
+              contains(github.ref_name, 'main') ||
+              contains(github.ref_name, 'runci/')
+            )
+          ) ||
+          github.event_name == 'schedule' ||
+          (
+            github.event_name == 'workflow_dispatch' &&
+            github.event.inputs.run-integration-tests == 'yes'
+          )
+        )
+      run: |
+        ./.github/workflows/build-test mypy integration
       env:
         PYTKET_RUN_REMOTE_TESTS: 1
-    - name: Build and test including remote checks (3.10) nomypy
-      if:  (matrix.os != 'macos-12') && (github.event_name == 'push' || (github.event_name == 'pull_request' &&  github.event.pull_request.head.repo.full_name == github.repository) || github.event_name == 'release' || github.event_name == 'schedule')
+    - name: Build and test (3.10) nomypy
+      if: |
+        matrix.os != 'macos-12' &&
+        (
+          github.event_name == 'push' ||
+          (
+            github.event_name == 'pull_request' &&
+            github.event.pull_request.head.repo.full_name == github.repository
+          ) ||
+          github.event_name == 'release' ||
+          (
+            github.event_name == 'workflow_dispatch' &&
+            github.event.inputs.run-integration-tests == 'no'
+          )
+        )
       shell: bash
       run: |
         ./.github/workflows/build-test nomypy
+    - name: Build and test including integration (3.10) nomypy
+      if: |
+        matrix.os != 'macos-12' &&
+        (
+          (
+            github.event_name == 'push' &&
+            (
+              contains(github.ref_name, 'main') ||
+              contains(github.ref_name, 'runci/')
+            )
+          ) ||
+          github.event_name == 'schedule' ||
+          (
+            github.event_name == 'workflow_dispatch' &&
+            github.event.inputs.run-integration-tests == 'yes'
+          )
+        )
+      shell: bash
+      run: |
+        ./.github/workflows/build-test nomypy integration
       env:
         PYTKET_RUN_REMOTE_TESTS: 1
     - name: Set up Python 3.11
-      if: github.event_name == 'push' || github.event_name == 'pull_request' || github.event_name == 'schedule'
+      if: |
+        github.event_name == 'push' ||
+        github.event_name == 'pull_request' ||
+        github.event_name == 'schedule' ||
+        github.event_name == 'workflow_dispatch'
       uses: actions/setup-python@v4
       with:
         python-version: '3.11'
     - name: Build and test (3.11)
-      if: github.event_name == 'push' || github.event_name == 'pull_request' || github.event_name == 'schedule'
+      if: |
+        github.event_name == 'push' ||
+        github.event_name == 'pull_request' ||
+        github.event_name == 'schedule' ||
+        github.event_name == 'workflow_dispatch'
       shell: bash
       run: |
         ./.github/workflows/build-test nomypy
